canvas correctly loads saved model, fix activity startup errors
I updated the python canvas to correctly pass saved models to
the engine, fixing the exception that was causing the canvas
to bail out of loading models early.

Moved the connections for the model and graph buttons to after
both objects were created, so that the toggle callback always
has both widgets available to make comparisons to.