changes to get opensim-gtk running on the mac
I've been able to pull together opensim-gtk on the mac.
After some of the challenges with compiling all the needed parts
of the pygtk toolchain, it required some conditional imports
of widgets originally from sugar, and the addition of some lines
to the search path for icons.  oh, and the use of png icons
instead of svg.  not sure why that is broken.